Added best-effort implementations of skipNBytes(long) in ServletInputStream delegates using reflection for Java 12 and later.
Any instances of these streams will not be referenced by their actual implementation classes but the ServletInputStream base class instead, because this is what the servlet APIs provide. Therefore, libraries / frameworks requiring Java 12 or later might call the method, and in this case the reflection based code will be able to invoke the skipNBytes method on the delegate stream. However, libraries / frameworks compiled against older Java versions will not be able to call the (in that case non-functioning) skipNBytes method, because it was never available in the corresponding servlet APIs at compilation time.